grafana/requirements.txt
docker==7.1.0
psycopg2-binary==2.9.9
pyyaml==6.0.2
kubernetes==30.1.0
ansible-runner==2.4.0